+++ tests/Unit/Traits/HasGeolocationPreferencesTest.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Bkhim\Geolocation\Tests\Unit\Traits;
-
-use Bkhim\Geolocation\Traits\HasGeolocationPreferences;
-use Illuminate\Support\Facades\Config;
-
-// All test methods removed as they require database interaction